⚡ updates

+++ SurvivorsLitClub/src/main/java/com/xoraus/SurvivorsLitClub/book/BookRepository.java
@@ -15,5 +15,5 @@ public interface BookRepository extends JpaRepository<Book, Integer>, JpaSpecifi
             AND book.owner.id != :userId
 """
     )
-    Page<Book> findAllDisplayableBooks(Pageable pageable, Long id);
+    Page<Book> findAllDisplayableBooks(Pageable pageable, Integer id);
 }
